echo_server/pyproject.toml
[project]
name = "echo-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]
